# implementation/virt-reader/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AJIN RFID Tracking API",
        "description": "RFID 기반 물류 추적 시스템 API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/reader-locations": {
            "get": {
                "tags": [
                    "Reader Locations"
                ],
                "summary": "List Reader Locations",
                "description": "리더기 위치 목록 조회 (페이지네이션 지원) (권한: reader_locations:read)\n\n- is_registered=true: 공정 매핑된 리더기만\n- is_registered=false: 미등록(공정 미매핑) 리더기만\n- 파라미터 없음: 전체 조회\n- process_id: 특정 공정에 속한 리더기만 조회",
                "operationId": "list_reader_locations_api_v1_reader_locations_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 20,
                            "title": "Per Page"
                        }
                    },
                    {
                        "name": "is_registered",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "boolean"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Is Registered"
                        }
                    },
                    {
                        "name": "process_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "필터링할 공정 ID",
                            "title": "Process Id"
                        },
                        "description": "필터링할 공정 ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReaderLocationListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reader Locations"
                ],
                "summary": "Create Reader Location",
                "description": "리더기 위치 등록 (권한: reader_locations:write)",
                "operationId": "create_reader_location_api_v1_reader_locations_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReaderLocationCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReaderLocationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reader-locations/{id}": {
            "get": {
                "tags": [
                    "Reader Locations"
                ],
                "summary": "Get Reader Location",
                "description": "리더기 위치 상세 조회 (권한: reader_locations:read)",
                "operationId": "get_reader_location_api_v1_reader_locations__id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReaderLocationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Reader Locations"
                ],
                "summary": "Update Reader Location",
                "description": "리더기 위치 수정 (권한: reader_locations:write)",
                "operationId": "update_reader_location_api_v1_reader_locations__id__put",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReaderLocationUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReaderLocationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reader Locations"
                ],
                "summary": "Delete Reader Location",
                "description": "리더기 위치 삭제 (권한: reader_locations:write)",
                "operationId": "delete_reader_location_api_v1_reader_locations__id__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reader-locations/port/{port_name}": {
            "get": {
                "tags": [
                    "Reader Locations"
                ],
                "summary": "Get Reader Location By Port",
                "description": "포트명으로 리더기 위치 조회 (권한: reader_locations:read)",
                "operationId": "get_reader_location_by_port_api_v1_reader_locations_port__port_name__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "port_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Port Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReaderLocationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reader-locations/{id}/register": {
            "put": {
                "tags": [
                    "Reader Locations"
                ],
                "summary": "Register Reader Location",
                "description": "리더기 등록 (공정 매핑) (권한: reader_locations:write)",
                "operationId": "register_reader_location_api_v1_reader_locations__id__register_put",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReaderLocationUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReaderLocationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reader-locations/test-connection": {
            "post": {
                "tags": [
                    "Reader Locations"
                ],
                "summary": "Test Reader Connection",
                "description": "RFID 리더기 연결 테스트\n\n현재는 시뮬레이션 모드로, 실제 하드웨어 연결을 테스트하지 않고\n포트명 형식만 검증합니다.",
                "operationId": "test_reader_connection_api_v1_reader_locations_test_connection_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "additionalProperties": true,
                                "type": "object",
                                "title": "Data"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dashboard/readers": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get Reader Status",
                "description": "리더기 상태 조회 (권한: dashboard:read)\n\n모든 등록된 리더기의 연결 상태와 마지막 스캔 시간을 반환합니다.\nlast_scan_time은 rfid_reader_locations 테이블에 직접 저장되어 조인 없이 빠르게 조회됩니다.",
                "operationId": "get_reader_status_api_v1_dashboard_readers_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReaderStatusList"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ReaderLocationCreate": {
                "properties": {
                    "port_name": {
                        "type": "string",
                        "title": "Port Name",
                        "description": "포트 이름 (COM3, 192.168.1.100:9001)"
                    },
                    "process_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Process Id",
                        "description": "공정 ID (미등록 시 None)"
                    },
                    "location_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Location Type",
                        "description": "위치 타입 (IN, OUT, HOLD, DEFECT, FINISH, RETURN)"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description",
                        "description": "리더기 설명"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active",
                        "description": "활성 여부",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "port_name"
                ],
                "title": "ReaderLocationCreate"
            },
            "ReaderLocationListResponse": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ReaderLocationResponse"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page"
                    },
                    "per_page": {
                        "type": "integer",
                        "title": "Per Page"
                    },
                    "pages": {
                        "type": "integer",
                        "title": "Pages"
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "total",
                    "page",
                    "per_page",
                    "pages"
                ],
                "title": "ReaderLocationListResponse"
            },
            "ReaderLocationResponse": {
                "properties": {
                    "created_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    },
                    "port_name": {
                        "type": "string",
                        "title": "Port Name",
                        "description": "포트 이름 (COM3, 192.168.1.100:9001)"
                    },
                    "process_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Process Id",
                        "description": "공정 ID (미등록 시 None)"
                    },
                    "location_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Location Type",
                        "description": "위치 타입 (IN, OUT, HOLD, DEFECT, FINISH, RETURN)"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description",
                        "description": "리더기 설명"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active",
                        "description": "활성 여부",
                        "default": true
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "display_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Display Name"
                    },
                    "process_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Process Name"
                    },
                    "process_code": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Process Code"
                    }
                },
                "type": "object",
                "required": [
                    "port_name",
                    "id"
                ],
                "title": "ReaderLocationResponse"
            },
...
            "ReaderStatusList": {
                "properties": {
                    "readers": {
                        "items": {
                            "$ref": "#/components/schemas/ReaderStatus"
                        },
                        "type": "array",
                        "title": "Readers"
                    }
                },
                "type": "object",
                "required": [
                    "readers"
                ],
                "title": "ReaderStatusList",
                "description": "리더기 상태 목록"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    }
}